requirements.txt
datasets
huggingface_hub
mteb
openai
python-dotenv
sentence-transformers
transformers